Fix pointer position handling
The coordinates provided in DXGKARG_SETPOINTERPOSITION when Windows
calls DxgkddiSetpointerposition are the coordinates of the upper left-
hand corner of the pointer, not of the pointer hotspot. (As far as I can
tell from the API documentation, the old XDDM provided the coordinates
of the hotspot.) The QXL device expects to receive the hotspot position.
This patch keeps track of the hotspot position for the current cursor so
that the driver can provide the correct coordinates to the QXL device.