Remove AppVeyor workaround
It appears that, at one point, the WDF headers caused issues for non-WDF
projects, so AppVeyor renamed the wdf directory to 00wdf. Before running
a build, we needed to move the directory to the right place. AppVeyor no
longer renames the wdf directory, so the workaround now causes an error
because the 00wdf directory does not exist.